+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Refactored the create or update access dialog component to standalone
 - Upgraded `envalid` from version `8.0.0` to `8.1.0`
 
 ### Fixed